autoresearch-inspired loop

+++ autoresearch_pg/candidates/mut_systems_float32_test/train_gpt_mlx.py
@@ -30,7 +30,7 @@ from mlx.utils import tree_flatten, tree_unflatten
 # SHARD FORMAT + COMPUTE DTYPE
 # ==============================================================================
 
-COMPUTE_DTYPE = mx.bfloat16
+COMPUTE_DTYPE = mx.float32
 
 # ==============================================================================
 # HYPERPARAMETERS
